+++ lib/Method.bs.js
@@ -1,0 +1,67 @@
+// Generated by BUCKLESCRIPT, PLEASE EDIT WITH CARE
+'use strict';
+
+
+function methodString(method_) {
+  switch (method_) {
+    case /* GET */0 :
+        return "GET";
+    case /* POST */1 :
+        return "POST";
+    case /* PUT */2 :
+        return "PUT";
+    case /* UPDATE */3 :
+        return "UPDATE";
+    case /* DELETE */4 :
+        return "DELETE";
+    case /* HEAD */5 :
+        return "HEAD";
+    case /* OPTION */6 :
+        return "OPTION";
+    case /* CONNECT */7 :
+        return "CONNECT";
+    case /* TRACE */8 :
+        return "TRACE";
+    case /* PATCH */9 :
+        return "PATCH";
+    
+  }
+}
+
+function getMethod(method_) {
+  if (method_ >= 492530731) {
+    if (method_ !== 801894688) {
+      if (method_ >= 965227017) {
+        if (method_ >= 994393768) {
+          return /* PATCH */9;
+        } else {
+          return /* UPDATE */3;
+        }
+      } else if (method_ >= 891112544) {
+        return /* POST */1;
+      } else {
+        return /* DELETE */4;
+      }
+    } else {
+      return /* HEAD */5;
+    }
+  } else if (method_ !== 3546230) {
+    if (method_ >= 336447077) {
+      if (method_ >= 357830602) {
+        return /* CONNECT */7;
+      } else {
+        return /* TRACE */8;
+      }
+    } else if (method_ >= 3997359) {
+      return /* PUT */2;
+    } else {
+      return /* OPTION */6;
+    }
+  } else {
+    return /* GET */0;
+  }
+}
+
+exports.methodString = methodString;
+exports.getMethod = getMethod;
+/* No side effect */
