+++ lib/ResponseNotFound.bs.js
@@ -1,0 +1,11 @@
+// Generated by BUCKLESCRIPT, PLEASE EDIT WITH CARE
+'use strict';
+
+var Rest = require("./Rest.bs.js");
+
+function run(param, response) {
+  return Rest.bodyResponse(response, 404, "File not Found");
+}
+
+exports.run = run;
+/* No side effect */
