+++ lib/ResponseNotFound.re
@@ -1,0 +1,3 @@
+let run = (_, response) => {
+    Rest.bodyResponse(response, 404, "File not Found")
+}
