+++ lib/Routers.bs.js
@@ -1,0 +1,22 @@
+// Generated by BUCKLESCRIPT, PLEASE EDIT WITH CARE
+'use strict';
+
+var Http = require("bs-node-http/src/Http.js");
+var Curry = require("bs-platform/lib/js/curry.js");
+var Method = require("./Method.bs.js");
+var Server = require("./Server.bs.js");
+var ResponseNotFound = require("./ResponseNotFound.bs.js");
+
+function start(request, response, routers) {
+  var methodOfRequest = Method.getMethod(Http.ClientRequest.getMethod(request));
+  var uriRequest = Http.ClientRequest.getUrl(request);
+  var methodSelected = Server.methodSelected(methodOfRequest, uriRequest, routers);
+  if (methodSelected !== undefined) {
+    return Curry._2(methodSelected.action, request, response);
+  } else {
+    return ResponseNotFound.run(request, response);
+  }
+}
+
+exports.start = start;
+/* No side effect */
