+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "reason-http-node",
-  "version": "0.0.6",
+  "version": "0.0.7",
   "scripts": {
     "dev": "nodemon lib/sample/App.bs.js",
     "dev:re": "bsb -make-world -w"
